feat: Refactor style creation using StyleFactory
The code changes in this commit refactor the creation of toastification styles by introducing a `StyleFactory` class. This class provides a static method `createStyle` that takes a `ToastificationStyle` and a `ToastificationType` as parameters and returns the corresponding built-in style.

This refactoring improves code organization and readability by separating the style creation logic from the `BuiltInStyle` class. It also makes it easier to add new styles in the future.

Recent user commits:
- refactor: add reflectable, make reflector and separate toastification style
- refactor: Move constants values to separate file
- refactor: Move constants values to separate file
- Add Toastification type as a class instead of enum
- refactor: replace row in built_in widgets instead of row
- feat: add show icon state to preview
- fix: fix surface color issues
- feat: update fvm version
- Merge pull request #2 from alirezat66/feature/remove_icon_posibility
- change toast_detail_ui_state